refactor: Split render pipeline interface into runtime and editor.
