+++ Assets/link.xml
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<linker>
+	 <assembly fullname="VisualPinball.Engine.Unity.Hdrp" preserve="all" />
+</linker>
